# docs/contracts/artifacts-v1.openapi.json
{
  "components": {
    "schemas": {
      "ArchiveImport": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "format": "uuid",
            "title": "Artifact Id",
            "type": "string"
          },
          "license": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "title": "License",
            "type": "string"
          },
          "license_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Url"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "name",
          "artifact_id",
          "license"
        ],
        "title": "ArchiveImport",
        "type": "object"
      },
      "ArtifactResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sha256",
          "size_bytes",
          "format",
          "created_at"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
      "CatalogEntry": {
        "additionalProperties": false,
        "properties": {
          "architecture": {
            "title": "Architecture",
            "type": "string"
          },
          "context_length": {
            "exclusiveMinimum": 0.0,
            "title": "Context Length",
            "type": "integer"
          },
          "gated": {
            "const": false,
            "title": "Gated",
            "type": "boolean"
          },
          "key": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,199}$",
            "title": "Key",
            "type": "string"
          },
          "license": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "title": "License",
            "type": "string"
          },
          "license_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "License Url",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Name",
            "type": "string"
          },
          "parameter_count": {
            "exclusiveMinimum": 0.0,
            "title": "Parameter Count",
            "type": "integer"
          },
          "quantization_support": {
            "items": {
              "type": "string"
            },
            "title": "Quantization Support",
            "type": "array"
          },
          "repository": {
            "pattern": "^[A-Za-z0-9_][A-Za-z0-9_.-]{0,95}/[A-Za-z0-9_][A-Za-z0-9_.-]{0,95}$",
            "title": "Repository",
            "type": "string"
          },
          "revision": {
            "pattern": "^[0-9a-f]{40}$",
            "title": "Revision",
            "type": "string"
          },
          "soup_compatibility": {
            "title": "Soup Compatibility",
            "type": "string"
          },
          "supported_recipes": {
            "items": {
              "type": "string"
            },
            "title": "Supported Recipes",
            "type": "array"
          },
          "vllm_compatibility": {
            "title": "Vllm Compatibility",
            "type": "string"
          },
          "vram_guidance": {
            "title": "Vram Guidance",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "repository",
          "revision",
          "architecture",
          "parameter_count",
          "context_length",
          "license",
          "license_url",
          "gated",
          "soup_compatibility",
          "vllm_compatibility",
          "supported_recipes",
          "quantization_support",
          "vram_guidance"
        ],
        "title": "CatalogEntry",
        "type": "object"
      },
      "DownloadResponse": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_in"
        ],
        "title": "DownloadResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "HealthResponse",
        "type": "object"
      },
      "HuggingFaceImport": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "repository": {
            "pattern": "^[A-Za-z0-9_][A-Za-z0-9_.-]{0,95}/[A-Za-z0-9_][A-Za-z0-9_.-]{0,95}$",
            "title": "Repository",
            "type": "string"
          },
          "revision": {
            "default": "main",
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_][A-Za-z0-9_./-]*$",
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "name",
          "repository"
        ],
        "title": "HuggingFaceImport",
        "type": "object"
      },
      "ModelView": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionView"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "versions"
        ],
        "title": "ModelView",
        "type": "object"
      },
      "SourceView": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "catalog_entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Entry Id"
          },
          "gated": {
            "title": "Gated",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "license_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Url"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "required": [
          "kind",
          "location",
          "revision",
          "license",
          "license_url",
          "gated",
          "notes",
          "artifact_id",
          "catalog_entry_id"
        ],
        "title": "SourceView",
        "type": "object"
      },
      "UploadRequest": {
        "additionalProperties": false,
        "properties": {
          "content_type": {
            "default": "application/octet-stream",
            "maxLength": 255,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9!#$&^_.+-]+/[a-zA-Z0-9!#$&^_.+-]+$",
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "maximum": 5368709120.0,
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "size_bytes"
        ],
        "title": "UploadRequest",
        "type": "object"
      },
      "UploadResponse": {
        "properties": {
          "complete_by": {
            "format": "date-time",
            "title": "Complete By",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "upload_id": {
            "format": "uuid",
            "title": "Upload Id",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "upload_id",
          "url",
          "headers",
          "expires_at",
          "complete_by"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionView": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceView"
            },
            "title": "Sources",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "sources"
        ],
        "title": "VersionView",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "KitchenSoup",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/artifact-uploads": {
      "post": {
        "operationId": "initiate_upload_api_v1_artifact_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Initiate Upload",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifact-uploads/{upload_id}/complete": {
      "post": {
        "operationId": "complete_upload_api_v1_artifact_uploads__upload_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Upload",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/download": {
      "post": {
        "operationId": "download_artifact_api_v1_artifacts__artifact_id__download_post",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/model-catalog": {
      "get": {
        "operationId": "catalog_api_v1_model_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogEntry"
                  },
                  "title": "Response Catalog Api V1 Model Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Catalog",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/model-catalog/{key}/register": {
      "post": {
        "operationId": "choose_api_v1_model_catalog__key__register_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Choose",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/model-imports/archive": {
      "post": {
        "operationId": "import_archive_api_v1_model_imports_archive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Archive",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/model-imports/huggingface": {
      "post": {
        "operationId": "import_huggingface_api_v1_model_imports_huggingface_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HuggingFaceImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Huggingface",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/models": {
      "get": {
        "operationId": "list_models_api_v1_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelView"
                  },
                  "title": "Response List Models Api V1 Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Models",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/models/{model_id}": {
      "get": {
        "operationId": "get_model_api_v1_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model",
        "tags": [
          "models"
        ]
      }
    },
    "/healthz": {
      "get": {
        "description": "Report process liveness; no infrastructure dependencies are checked.",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    }
  }
}